% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_ext.R
\name{ggsave_ext}
\alias{ggsave_ext}
\alias{ggsave_social}
\title{Save a ggplot2 plot to file and update file EXIF metadata}
\usage{
ggsave_ext(
  plot = ggplot2::last_plot(),
  name = NULL,
  label = NULL,
  prefix = NULL,
  postfix = NULL,
  filename = NULL,
  device = NULL,
  filetype = NULL,
  path = NULL,
  paper = NULL,
  orientation = "portrait",
  width = NULL,
  height = NULL,
  units = "in",
  scale = 1,
  dpi = 300,
  bgcolor = NULL,
  exif = FALSE,
  title = NULL,
  author = NULL,
  keywords = NULL,
  args = NULL,
  ...
)

ggsave_social(
  plot = ggplot2::last_plot(),
  image = "Instagram post",
  platform = NULL,
  format = NULL,
  orientation = NULL,
  name = NULL,
  filename = NULL,
  filetype = "jpeg",
  dpi = 72,
  width = 1080,
  height = 1080,
  units = "px",
  ...
)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{name}{Plot name, used to create filename (if filename is \code{NULL}) using
\code{\link[sfext:make_filename]{sfext::make_filename()}}}

\item{filename}{File name to create on disk.}

\item{device}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{filetype}{The file extension or file type; defaults to \code{NULL}.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{paper}{Paper matching name from \code{paper_sizes} (e.g. "letter"). Not case
sensitive.}

\item{orientation}{Page orientation ("portrait", "landscape", or "square").}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{scale}{Multiplicative scaling factor.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{bgcolor}{Background color to optionally override \code{plot.background}
theme element.}

\item{exif}{If \code{TRUE}, the EXIF metadata for the exported file is updated
with the exifr package; defaults to \code{FALSE}.}

\item{title}{Title to add to file metadata with exiftoolr, Default: \code{NULL}.}

\item{author}{Author to add to file metadata with exiftoolr, Default: \code{NULL}.}

\item{keywords}{Keyword(s) added to file metadata with with exiftoolr,
Default: \code{NULL}.}

\item{args}{Alternate arguments passed to \code{\link[exiftoolr:exif_call]{exiftoolr::exif_call()}}. If args
is not \code{NULL}, title, author, date, and keywords are ignored; defaults to
\code{NULL}.}

\item{...}{Additional parameters passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\description{
Save a plot or map then update the EXIF metadata for the title, author, and
create data. \code{\link[=ggsave_ext]{ggsave_ext()}} also supports creating a file name based on a
sentence case name with spaces (e.g. "Baltimore city map") and appending a
label (e.g. "baltcity") as a prefix to the output file name.
}
\examples{
\dontrun{
  ggplot2::ggplot() +
    layer_location_data(
      data = system.file("shape/nc.shp", package = "sf")
    )

  ggsave_ext(
    name = "counties",
    label = "North Carolina",
    device = "pdf",
    paper = "letter"
    )
}
}
\seealso{
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
}
