% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_arrow.R
\name{layer_arrow}
\alias{layer_arrow}
\title{Create a layer with an arrow or segment from and to specified locations}
\usage{
layer_arrow(
  mapping = NULL,
  data = NULL,
  crs = NULL,
  from,
  to,
  geom = "segment",
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping overwritten with x, y, xend, and yend values
based on provided from and to parameters.}

\item{data}{Required if from or to are character vectors to \code{\link[sfext:as_xy]{sfext::as_xy()}}}

\item{crs}{A character or numeric reference to a coordinate reference system
supported by \code{\link[sf:st_crs]{sf::st_crs()}} or another  \code{sf}, \code{sfc}, or \code{bbox} object that
is used to provide crs.}

\item{from, to}{Required. Passed to x parameter of \code{\link[sfext:as_xy]{sfext::as_xy()}} (using \code{nm = c("xend", "yend")}) for the to parameter.}

\item{geom}{Character string for geom to use  c("segment", "curve",
"arrowsegment", "diagonal0", "link") or a geom function.}

\item{...}{Additional parameters passed to function specified by geom
paramter.}
}
\description{
A wrapper for \link[ggplot2:geom_segment]{ggplot2::geom_segment}, \code{\link[ggplot2:geom_curve]{ggplot2::geom_curve()}},
\code{\link[ggarchery:geom_arrowsegment]{ggarchery::geom_arrowsegment()}}, \code{\link[ggforce:geom_diagonal0]{ggforce::geom_diagonal0()}},
\code{\link[ggforce:geom_link]{ggforce::geom_link()}} that makes it easier to specify the start and end of
the segment using any object supported by the \code{\link[sfext:as_xy]{sfext::as_xy()}} function.
}
\examples{
library(ggplot2)

nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))

nc_map <- ggplot(data = nc) +
  geom_sf()

nc_map +
  layer_arrow(
    data = nc,
    from = c("xmin", "ymin"),
    to = c("xmid", "ymax"),
  )

nc_map +
  layer_arrow(
    data = nc,
    from = c("xmin", "ymin"),
    to = c("xmid", "ymax"),
    geom = "curve",
    curvature = 0.25
  )

nc_map +
  layer_arrow(
    data = nc,
    from = c("xmax", "ymin"),
    to = c("xmid", "ymax"),
    geom = "arrowsegment"
  )
}
\seealso{
\code{\link[ggplot2:reexports]{ggplot2::reexports()}}, \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}, \code{\link[ggplot2:aes]{ggplot2::aes()}}
}
