% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_image_path.R
\name{layer_image_path}
\alias{layer_image_path}
\title{Use ggpath to create a layer with images at locations}
\usage{
layer_image_path(
  data = NULL,
  path = NULL,
  path_col = "path",
  width = 0.1,
  crs = getOption("maplayer.crs", 3857),
  segment_params = NULL,
  neatline = FALSE,
  basemap = FALSE,
  ...
)
}
\arguments{
\item{data}{A \code{sf} object with a column containing file paths that has the
same name as the path_col argument. Optional if "path" is provided for
\code{\link[sfext:read_sf_exif]{sfext::read_sf_exif()}}. Required if path is \code{NULL}. If path is provided,
data is ignored.}

\item{path}{A path to folder of one or more files with EXIF location
metadata.}

\item{path_col}{Column name with file paths from data. Defaults to "path"
(path name used by data returned from \code{\link[sfext:read_sf_exif]{sfext::read_sf_exif()}})}

\item{width}{Width of the image in npc (Normalised Parent Coordinates) passed
to \code{\link[ggpath:geom_from_path]{ggpath::geom_from_path()}}; defaults to 0.1.}

\item{crs}{The coordinate reference system (CRS) into which all data should
be projected before plotting. If not specified, will use the CRS defined
in the first sf layer of the plot.}

\item{segment_params}{Not implemented: parameters to define segments
connecting images to the location.}

\item{neatline}{A logical object, \code{CoordSf} object, or a list containing a
\code{CoordSf} object (typically from \code{\link[=layer_neatline]{layer_neatline()}}) added to layer by
\code{\link[=set_neatline]{set_neatline()}}.
\itemize{
\item If logical and \code{TRUE}, add a neatline layer using data, crs and any
additional parameters passed to ... If logical and \code{FALSE}, return x as is.
\item If object from \code{\link[=layer_neatline]{layer_neatline()}}, add it as is.
}}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \link[ggplot2:ggplot]{ggplot2::ggplot}. If \code{FALSE}, return x
as is.

If a \strong{ggplot}, add x to basemap object.}

\item{...}{Additional EXIF tags to pass to \link[exiftoolr:exif_read]{exiftoolr::exif_read}}
}
\description{
Use the \code{\link[ggpath:geom_from_path]{ggpath::geom_from_path()}} function with \code{\link[sfext:read_sf_exif]{sfext::read_sf_exif()}} and
\code{stat = "sf_coordinates"} to create a layer showing images at locations.
}
