% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_location_context.R
\name{layer_location_context}
\alias{layer_location_context}
\alias{layer_show_context}
\title{Create a layer showing a location and related context}
\usage{
layer_location_context(
  data = NULL,
  location = NULL,
  fill = "gray70",
  color = "black",
  context = NULL,
  context_params = list(fill = "white", color = "black", alpha = 1, ...),
  layer_between = NULL,
  crs = getOption("maplayer.crs", default = 3857),
  neatline = TRUE,
  basemap = FALSE,
  ...
)
}
\arguments{
\item{data, location}{data and location can be a either sf object for the
location or a ggplot layer representing the location. location can also be
a formula or a function used to subset context (requires context to be an
sf object). If data and context are both provided and both are sf objects,
data is ignored.}

\item{fill, color}{Fill and color aesthetics for location data}

\item{context}{A sf object for context area or a ggplot layer representing
the context.}

\item{context_params}{list with parameters for context layer; defaults to
list(fill = "white", color = "black", alpha = 1, ...).}

\item{layer_between}{An optional ggplot2 layer to insert between the context
layer and the location layer.}

\item{crs}{Coordinate reference system to return.}

\item{neatline}{If \code{TRUE}, add a neatline layer to the returned layer.
Requires that data be an sf object and not inherited.}

\item{...}{Additional parameters passed to \link{layer_location_data} for location}
}
\description{
Intended for use with \link{make_inset_map} function.
}
