% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_scaled.R
\name{layer_scaled}
\alias{layer_scaled}
\title{Create a ggplot2 layer scaled to a paper and orientation for a location}
\usage{
layer_scaled(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = getOption("maplayer.crs", default = 3857),
  scale = NULL,
  paper = NULL,
  orientation = NULL,
  clip = FALSE
)
}
\arguments{
\item{data}{A \code{sf}, \code{sfc}, or \code{bbox} class object.}

\item{dist}{distance to convert. If paper is provided, dist is optional and
paper width and height are used as dist.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Coordinate reference system of bounding box to return; defaults to
\code{NULL} which maintains the crs of the input object.}

\item{scale}{Scale name from \code{standard_scales[["scale"]]}.}

\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Orientation "portrait", "landscape", or "square", Default:
'portrait'.}

\item{clip}{If \code{TRUE}, create scaled layer even if the data is cut off;
defaults to \code{FALSE}.}
}
\description{
Uses \link{layer_neatline}, \link[sfext:standard_scales]{sfext::standard_scales}, and
\link[sfext:convert_dist_scale]{sfext::convert_dist_scale}.
}
\seealso{
Other layer:
\code{\link[=layer_frame]{layer_frame()}},
\code{\link[=layer_location_data]{layer_location_data()}},
\code{\link[=layer_neatline]{layer_neatline()}}
}
\concept{layer}
