% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_atlas.R
\name{make_atlas}
\alias{make_atlas}
\title{Convert a set of ggplot2 maps to an atlas}
\usage{
make_atlas(
  plots,
  dims = NULL,
  ncol = NULL,
  nrow = NULL,
  page = "letter",
  orientation = "portrait",
  save = FALSE,
  filename = NULL,
  ...
)
}
\arguments{
\item{plots}{A list of ggplot2 maps to assemble into a set of sheet maps in
an atlas format.}

\item{dims}{Optional. Plot dimensions. Ignored if ncol and nrow are supplied.
Otherwise, if \code{NULL} (default), dims are inferred based on the dimensions
of the first plot in plots.}

\item{ncol, nrow}{The dimensions of the grid to create.  If both are \code{NULL},
dims will be used or dims will be determined based on the plot dimensions.}

\item{page}{Used by \code{\link[papersize:get_page_dims]{get_page_dims()}}, page is either a character vector
passed to the name parameter of \code{\link[papersize:get_page_size]{get_page_size()}}, a data.frame with column
names matching the cols parameter, or a length 2 numeric vector with the
page width and height.}

\item{orientation}{Page orientation, Default: \code{NULL}. Supported options are
"portrait", "landscape", or "square".}

\item{save}{If \code{TRUE}, save atlas plots to files using
\code{\link[papersize:map_ggsave_ext]{papersize::map_ggsave_ext()}} Default: FALSE}

\item{filename}{File name to create on disk.}

\item{...}{
  Arguments passed on to \code{\link[papersize:map_ggsave_ext]{papersize::map_ggsave_ext}}
  \describe{
    \item{\code{name}}{Plot name, used to create filename (if filename is \code{NULL}) using
\code{\link[filenamr:make_filename]{filenamr::make_filename()}}}
    \item{\code{label}}{Label to combine with name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}. The label is designed to identify the area
or other shared characteristics across multiple data files, maps, or plots.
label is ignored if name is NULL or if name includes a file extension.}
    \item{\code{prefix}}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}
    \item{\code{postfix}}{File name postfix; defaults to \code{NULL}.}
    \item{\code{increment}}{If \code{TRUE}, increment digits in string by 1. If numeric,
increment digits in string by value. If \code{NULL}, 0, or if no digits are
present in string, return string as is.}
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}
    \item{\code{fileext}}{File type or extension. Optional if filename or path include a
file extension.}
    \item{\code{filetype}}{File type (used if fileext is \code{NULL}).}
    \item{\code{path}}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}
    \item{\code{paper}}{Paper matching name from \code{paper_sizes} (e.g. "letter"). Not case
sensitive.}
    \item{\code{width,height}}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}
    \item{\code{asp}}{Numeric aspect ratio used to determine width or height if only one
of the two arguments is provided; defaults to \code{NULL}.}
    \item{\code{units}}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{dpi}}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Only applies when converting pixel units,
as is typical for raster output types.}
    \item{\code{bgcolor}}{Background color to optionally override \code{plot.background}
theme element.}
    \item{\code{exif}}{If \code{TRUE}, the EXIF metadata for the exported file is updated
with the exifr package; defaults to \code{FALSE}.}
    \item{\code{title}}{Title to add to file metadata with exiftoolr, Default: \code{NULL}.}
    \item{\code{author}}{Author to add to file metadata to the "Author" and
"XMP-dc:creator" tags. Default: \code{NULL}.}
    \item{\code{keywords}}{Keyword(s) added to file metadata to "IPTC:Keywords" and
"XMP-dc:Subject" tags. Defaults to \code{NULL}.}
    \item{\code{args}}{Alternate arguments passed to \code{\link[exiftoolr:exif_call]{exiftoolr::exif_call()}}. Other tag
parameters are appended to args if they are not \code{NULL}.}
    \item{\code{overwrite}}{If \code{TRUE} (default), overwrite any existing file with the
same name or ask to overwrite if \code{ask = TRUE}. Passed to
\code{\link[filenamr:check_file_overwrite]{filenamr::check_file_overwrite()}}.}
    \item{\code{ask}}{If \code{TRUE}, ask before overwriting file with the same name.
Defaults to \code{FALSE}. Passed to \code{\link[filenamr:check_file_overwrite]{filenamr::check_file_overwrite()}}.}
    \item{\code{preview}}{If \code{TRUE}, open saved file in default system application.
Based on ggpreview from tjmisc package.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{quiet}}{If \code{TRUE} (default), suppress function messages.}
    \item{\code{image}}{Image name passed to name parameter of \code{\link[papersize:get_social_size]{get_social_size()}}.}
    \item{\code{platform}}{Social media platform, "Instagram", "Facebook", or "Twitter",
Default: \code{NULL}}
    \item{\code{format}}{Image format, "post", "story", or "cover", Default: \code{NULL}}
    \item{\code{single_file,onefile}}{If \code{TRUE}, use \code{\link[gridExtra:arrangeGrob]{gridExtra::arrangeGrob()}} to create an
arrangelist class object that \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} can save as a single
multi-page file. Note: this does not work with plots modified with
patchwork including inset maps created with the
\code{\link[maplayer:layer_inset]{maplayer::layer_inset()}} function.}
  }}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\code{\link[=make_atlas]{make_atlas()}} is a wrapper for \code{\link[papersize:page_layout]{papersize::page_layout()}} and
\code{\link[papersize:map_ggsave_ext]{papersize::map_ggsave_ext()}} with the intent for taking a list of maps into
a set of patchwork plots and optionally save plots to file. The function is
similar to \code{\link[papersize:make_contact_sheets]{papersize::make_contact_sheets()}}.
}
\details{
DETAILS
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))

plots <- lapply(
  dplyr::nest_by(nc, .by = NAME)[["data"]][1:4],
  function(x) {
    make_location_map(
      basemap = ggplot(),
      layer = layer_location(
        data = x,
        fill = "yellow",
        alpha = 0.5
      ),
      bg_layer = layer_location_data(
        data = nc,
        location = x,
        asp = 8.5 / 5.5,
        crop = FALSE
      ),
      neatline = layer_neatline(data = x, asp = 8.5 / 5.5),
      addon = labs_ext(caption = x$NAME)
    )
  }
)

make_atlas(
  plots = plots,
  page = "letter",
  nrow = 2,
  ncol = 1,
  save = FALSE
)
}
\keyword{internal}
