% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location_map.R
\name{make_basemap}
\alias{make_basemap}
\title{Create a base map by adding the object}
\usage{
make_basemap(x, basemap = FALSE)
}
\arguments{
\item{x}{A ggproto object or list of ggproto objects.}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \link[ggplot2:ggplot]{ggplot2::ggplot}. If \code{FALSE}, return x
as is.

If a \strong{ggplot}, add x to basemap object.}
}
\description{
Create a base map by adding the object
}
