% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_icons}
\alias{map_icons}
\title{Map icons}
\format{
A data frame with 1855 rows and 5 variables:
\describe{
\item{\code{name}}{Icon name}
\item{\code{url}}{Icon URL on GitHub repo}
\item{\code{style}}{Icon style (Weather Underground icons only)}
\item{\code{size}}{Icon width/height (pixels)}
\item{\code{repo}}{GitHub repository for icon collection}
}
}
\usage{
map_icons
}
\description{
An index of map icons from four sources:
\itemize{
\item \href{https://github.com/mapbox/maki}{mapbox/maki}
\item \href{https://github.com/ideditor/temaki}{ideditor/temaki}
\item \href{https://github.com/manifestinteractive/weather-underground-icons/}{manifestinteractive/weather-underground-icons}
\item \href{https://github.com/openstreetmap/map-icons/}{openstreetmap/map-icons}
\item \href{https://github.com/openstreetmap/lane-icons/}{openstreetmap/lane-icons}
\item \href{https://github.com/Esri/calcite-point-symbols}{Esri/calcite-point-symbols}
\item \href{https://github.com/nationalparkservice/symbol-library/}{NPMap Symbol Library}
}

Most of these icon sources use open licenses.
\href{https://github.com/mapbox/maki/blob/main/LICENSE.txt}{Maki},
\href{https://github.com/ideditor/temaki/blob/main/LICENSE.md}{Temaki}, and the
\href{https://github.com/openstreetmap/lane-icons/blob/master/LICENSE.md}{OSM lane icons}
all use a CC0 license. The Weather Underground Icons use an
\href{https://github.com/manifestinteractive/weather-underground-icons/blob/master/LICENSE}{MIT license}.
The OSM map icons use an unspecified PD style license.
The Calcite icons are \href{https://github.com/Esri/calcite-point-symbols#licensing}{available under the Esri Master License Agreement (MLA)}. The NPMap
Symbol Library is created by the National Park Service so is assumed to be a
\href{http://www.usa.gov/publicdomain/label/1.0/}{Public Domain work in the U.S.}.
}
\keyword{datasets}
