% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplayer-package.R
\docType{package}
\name{maplayer-package}
\alias{maplayer}
\alias{maplayer-package}
\title{maplayer: Make Map Layers With ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Make map-making with ggplot2 and sf more convenient with layers that subset by location.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/maplayer}
  \item \url{https://elipousson.github.io/maplayer/}
  \item Report bugs at \url{https://github.com/elipousson/maplayer/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]
}

}
\keyword{internal}
