% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot}
\alias{aes}
\alias{geom_sf}
\alias{geom_sf_text}
\alias{geom_sf_label}
\alias{get_last_plot}
\alias{margin}
\alias{guides}
\alias{ggsave_ext}
\alias{map_ggsave_ext}
\alias{ggsave_social}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:aes]{aes()}}, \code{\link[ggplot2:geom_sf]{geom_sf()}}, \code{\link[ggplot2:geom_sf_label]{geom_sf_label()}}, \code{\link[ggplot2:geom_sf_text]{geom_sf_text()}}, \code{\link[ggplot2:get_last_plot]{get_last_plot()}}, \code{\link[ggplot2:ggplot]{ggplot()}}, \code{\link[ggplot2:guides]{guides()}}, \code{\link[ggplot2:margin]{margin()}}}

  \item{papersize}{\code{\link[papersize:ggsave_ext]{ggsave_ext()}}, \code{\link[papersize:ggsave_social]{ggsave_social()}}, \code{\link[papersize:map_ggsave_ext]{map_ggsave_ext()}}}
}}

