% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ext.R
\name{theme_ext}
\alias{theme_ext}
\alias{theme_text}
\alias{theme_margin}
\alias{theme_legend}
\title{Modify the text, margins, or legend for a ggplot theme}
\usage{
theme_text(
  font_family = NULL,
  color = "black",
  geom_text = TRUE,
  hjust = NULL,
  vjust = NULL,
  method = NULL,
  ...
)

theme_margin(
  margin = "standard",
  paper = NULL,
  orientation = NULL,
  dist = NULL,
  unit = "in",
  block_width = NULL,
  header = 0,
  footer = 0,
  fill = NA,
  color = NA,
  linewidth = 0,
  size = NULL,
  method = NULL,
  ...
)

theme_legend(
  position = NULL,
  justification = NULL,
  margin = 8,
  unit = "pt",
  inset = TRUE,
  nudge_inset = 0.05,
  bgcolor = "white",
  title = list(face = "bold", align = 0),
  method = NULL,
  ...
)
}
\arguments{
\item{font_family}{Font family, Default: 'Helvetica' If \code{NULL}, font_family
is pulled from current set theme which is helpful for resetting all text
families to the theme.}

\item{color}{Color for text elements (passed to
\code{\link[ggplot2:element_text]{ggplot2::element_text()}} by theme_text), \code{plot.background}
(passed to \code{\link[ggplot2:element_rect]{ggplot2::element_rect()}} by theme_margin). Default:
\code{NA}.}

\item{geom_text}{If \code{TRUE}, update text family for
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}, \code{\link[ggplot2:geom_sf_text]{ggplot2::geom_sf_text()}},
\code{\link[ggplot2:geom_label]{ggplot2::geom_label()}}, and \code{\link[ggplot2:geom_sf_label]{ggplot2::geom_sf_label()}}
to match \code{font_family} and color. If \code{FALSE}, make no changes to the theme.
Default: \code{TRUE}.}

\item{hjust, vjust}{Horizontal and vertical justification.}

\item{method}{Method with name of the ggplot2 geom function to use for
modifying theme ("set", "update", or "replace"); defaults to \code{NULL}.}

\item{...}{Additional parameters passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{margin}{Margin distance, a margin style supported by \code{\link[=get_margin]{get_margin()}} or
a margin object; defaults to 10.}

\item{paper}{Paper, Default: 'letter'.}

\item{orientation}{Page orientation, Default: \code{NULL}. Supported options are
"portrait", "landscape", or "square".}

\item{dist}{Margin distance (single value used to all sides), Default: \code{NULL}}

\item{unit}{Legend margin units; defaults to 'pt'.}

\item{block_width}{Plot or map width in units. If \code{paper} and \code{block_width} are
provided, margins are half the distance between the two evenly distributed.
This sets the margin distance for height as well as width so does not work
well with header and footers and should be improved in the future.}

\item{header, footer}{Header and footer height in units; defaults to 0. Please
note: headers and footers are not currently supported for "px" units.}

\item{fill}{Fill for \code{plot.background} theme element passed to
\code{\link[ggplot2:element_rect]{ggplot2::element_rect()}} Default: \code{NA}.}

\item{linewidth, size}{Passed to linewidth of \code{\link[ggplot2:element_rect]{ggplot2::element_rect()}} to
define the plot.background theme element. linewidth is ignored if size is
provided.}

\item{position}{Legend position (“left”,“top”, “right”, “bottom”) or a
two-element numeric vector to set position using Normalized Parent
Coordinates ("npc"); defaults NULL}

\item{justification}{If \code{NULL}, justification is set to "center"; defaults to
\code{NULL}. Use justification to set legend position if \code{inset = FALSE}.
Supports "topleft", "bottomleft", "topright", or "bottomright" values.}

\item{inset}{If \code{TRUE} and position is "topleft", "bottomleft", "topright",
or "bottomright", place the legend in an inset position; defaults to
\code{TRUE}.}

\item{nudge_inset}{Position adjustment in "npc" units to use for inset
legends. Defaults to 0.05.}

\item{bgcolor}{Fill color for legend background; defaults to 'white'.}

\item{title}{Attributes to use for legend.title text (e.g. face and align).}
}
\description{
Helper functions for modifying a ggplot theme. The "replace" and "update"
options for the method parameter are not currently working; keeping method =
NULL or method = "set" is recommended.
}
\seealso{
\itemize{
\item \code{\link[ggplot2:theme]{ggplot2::theme()}}
\item \code{\link[ggplot2:margin]{ggplot2::margin()}}
\item \code{\link[ggplot2:theme_get]{ggplot2::theme_get()}}
\item \code{\link[ggplot2:update_geom_defaults]{ggplot2::update_geom_defaults()}}
}
}
