% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs_ext.R
\name{labs_ext}
\alias{labs_ext}
\title{Add labels to a ggplot2 plot or map}
\usage{
labs_ext(
  title = ggplot2::waiver(),
  subtitle = ggplot2::waiver(),
  caption = ggplot2::waiver(),
  tag = ggplot2::waiver(),
  alt = ggplot2::waiver(),
  fill = ggplot2::waiver(),
  color = ggplot2::waiver(),
  size = ggplot2::waiver(),
  shape = ggplot2::waiver(),
  alpha = ggplot2::waiver(),
  linewidth = ggplot2::waiver(),
  location = NULL,
  name_col = NULL,
  source = NULL,
  .na = "NA",
  .null = NULL,
  .envir = parent.frame(),
  use_md = FALSE,
  ...
)
}
\arguments{
\item{title}{The text for the title.}

\item{subtitle}{The text for the subtitle for the plot which will be
displayed below the title.}

\item{caption}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}

\item{tag}{The text for the tag label which will be displayed at the
top-left of the plot by default.}

\item{alt}{Text used for the generation of alt-text for the plot.}

\item{fill, color, size, shape, alpha, linewidth}{Optional labels for mapped
aesthetics. Defaults to \code{\link[ggplot2:waiver]{ggplot2::waiver()}}.}

\item{location}{sf or bbox object or character string}

\item{name_col}{Column name holding name or identifier for distinct places
within the simple feature collection provided to location. Not supported
for bbox objects.}

\item{source}{Data source(s). Not yet used or supported by function.}

\item{.na, .null}{Additional parameters passed to \code{\link[glue:glue]{glue::glue()}}}

\item{.envir}{Environment passed to \code{\link[glue:glue]{glue::glue()}}; defaults to
\code{\link[=parent.frame]{parent.frame()}}.}

\item{use_md}{If \code{TRUE}, use the ggtext::element_markdown() element for the
plot title, subtitle, and caption. Defaults to \code{FALSE}.}

\item{...}{A list of new name-value pairs. The name should be an aesthetic.}
}
\description{
A helper function that converts strings to glue strings for the title,
subtitle, and caption. In progress.
}
