% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_arrow.R
\name{layer_arrow}
\alias{layer_arrow}
\title{Create a layer with an arrow or segment from and to specified locations}
\usage{
layer_arrow(
  mapping = NULL,
  data = NULL,
  crs = NULL,
  from,
  to,
  geom = "segment",
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping overwritten with x, y, xend, and yend values
based on provided from and to parameters.}

\item{data}{Required if from or to are character vectors to \code{\link[sfext:as_xy]{sfext::as_xy()}}}

\item{crs}{A character or numeric reference to a coordinate reference system
supported by \code{\link[sf:st_crs]{sf::st_crs()}} or another  \code{sf}, \code{sfc}, or \code{bbox} object that
is used to provide crs.}

\item{from, to}{Required. Passed to x parameter of \code{\link[sfext:as_xy]{sfext::as_xy()}} (using \code{nm = c("xend", "yend")}) for the to parameter.}

\item{geom}{Character string for geom to use  c("segment", "curve",
"arrowsegment", "diagonal0", "link") or a geom function.}

\item{...}{Additional parameters passed to \code{\link[sfext:sf_bbox_point]{sf_bbox_point()}}, \code{\link[sfext:as_points]{as_points()}},
or \code{\link[sfext:as_sfc]{as_sfc()}}.}
}
\description{
A wrapper for \link[ggplot2:geom_segment]{ggplot2::geom_segment}, \code{\link[ggplot2:geom_segment]{ggplot2::geom_curve()}},
\code{\link[ggarchery:geom_arrowsegment]{ggarchery::geom_arrowsegment()}}, \code{\link[ggforce:geom_diagonal]{ggforce::geom_diagonal0()}},
\code{\link[ggforce:geom_link]{ggforce::geom_link()}} that makes it easier to specify the start and end of
the segment using any object supported by the \code{\link[sfext:as_xy]{sfext::as_xy()}} function.
}
\seealso{
\code{\link[ggplot2:reexports]{ggplot2::reexports()}}, \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}, \code{\link[ggplot2:aes]{ggplot2::aes()}}
}
