% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_frame.R
\name{layer_frame}
\alias{layer_frame}
\alias{make_frame}
\title{Create a frame layer around a simple feature object}
\usage{
layer_frame(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "meter",
  asp = NULL,
  style = "circle",
  scale = 1,
  rotate = 0,
  inscribed = FALSE,
  color = "black",
  linewidth = 0.5,
  linetype = "solid",
  fill = "white",
  neatline = TRUE,
  expand = TRUE,
  basemap = FALSE,
  union = TRUE,
  by_feature = FALSE,
  ...
)

make_frame(
  x,
  dist = NULL,
  diag_ratio = NULL,
  unit = "meter",
  asp = NULL,
  style = "circle",
  scale = 1,
  rotate = 0,
  inscribed = FALSE,
  dTolerance = 0,
  union = TRUE,
  by_feature = FALSE
)
}
\arguments{
\item{data, x}{A \code{sf}, \code{sfc}, or \code{bbox} object to create the frame around.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{style}{Style of framing shape to add, "circle", "square", "rect",
"buffer", or "none". If style is "buffer", the asp parameter is ignored. If
style is "none", the dist, diag_ratio, and asp parameters are ignored and
the input data is used as the frame.}

\item{scale}{numeric; scale factor, Default: 1}

\item{rotate}{numeric; degrees to rotate (-360 to 360), Default: 0}

\item{inscribed}{If \code{TRUE}, the returned geometry is inscribed within x, if
\code{FALSE} (default), the geometry is circumscribed.}

\item{fill, color, linewidth, linetype}{Fixed aesthetics for frame, passed to
\link{layer_location_data}.}

\item{neatline}{If \code{TRUE}, return a list of layers that includes a
\link{layer_neatline}}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.
Giving a logical vector will separately control the expansion for the four
directions (top, left, bottom and right). The \code{expand} argument will be
recycled to length 4 if necessary. Alternatively, can be a named logical
vector to control a single direction, e.g. \code{expand = c(bottom = FALSE)}.}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. If \code{FALSE}, return
x as is.

If a \strong{ggplot}, add x to basemap object.

If a \strong{ggproto} object (or list that contains a \strong{ggproto} object), add x
and basemap object to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{union}{If \code{TRUE}, pass data to \code{\link[sf:geos_combine]{sf::st_union()}} before buffering and
creating frame; defaults to \code{TRUE}.}

\item{by_feature}{If \code{TRUE}, create a frame around each feature. If \code{FALSE}
(default), union the provided features before creating a frame.}

\item{...}{
  Arguments passed on to \code{\link[=layer_location_data]{layer_location_data}}
  \describe{
    \item{\code{layer_fn}}{ggplot2 geom or custom function using lambda syntax. Use for
passing custom mapping functions to layer_location_data beyond the
supported geom options.}
    \item{\code{label_col}}{Column name or id for a column with the text or labels to
pass to any text geom.}
    \item{\code{smooth_params}}{Optional. Logical or a list of parameters passed to
\code{\link[smoothr:smooth]{smoothr::smooth()}}. If \code{TRUE}, apply \code{\link[smoothr:smooth]{smoothr::smooth()}} to location data
using default parameters. smooth_params is ignored if data is \code{NULL}
(inheriting data from ggplot).}
    \item{\code{shadow_params}}{Optional. Logical or a list of parameters passed to
\code{\link[ggfx:with_shadow]{ggfx::with_shadow()}}. If \code{TRUE}, apply \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}} to the layer
using default parameters. shadow_params is ignored if layer_fn is provided.}
    \item{\code{location}}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}}
    \item{\code{fileext,filetype}}{File extension or type to use if passing parameters
to \code{\link[sfext:read_sf_ext]{sfext::read_sf_download()}} or \code{\link[sfext:read_sf_ext]{sfext::read_sf_pkg()}} (required for
extdata and cached data).}
    \item{\code{fn}}{Function to apply to data after filtering by location but before
returning from function.}
    \item{\code{crop}}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}
    \item{\code{trim}}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}
    \item{\code{crs}}{Coordinate reference system to return.}
    \item{\code{mapping}}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}
  }}

\item{dTolerance}{numeric; tolerance parameter, specified for all or for each
feature geometry. If you run \code{st_simplify}, the input data is
specified with long-lat coordinates and \code{sf_use_s2()} returns
\code{TRUE}, then the value of \code{dTolerance} must be specified in
meters.}
}
\description{
Create a circle or square that can be used as a frame around a simple feature
object using fixed aesthetics for fill, color, size, and linetype. This
function is helpful for the background of an inset map intended for use with
\code{\link[=layer_inset]{layer_inset()}}.
}
\details{
The \code{\link[=make_frame]{make_frame()}} helper function calls \code{\link[sfext:st_misc]{sfext::st_circle()}} (if \code{style = "circle"}), \code{\link[sfext:st_square]{sfext::st_square()}}  (if \code{style = "square"}),
\code{\link[sfext:st_bbox_ext]{sfext::st_bbox_ext()}} (if \code{style = "rect"}), or \code{\link[sfext:st_buffer_ext]{sfext::st_buffer_ext()}} (if
\code{style = "none"}).

If neatline is \code{TRUE}, \code{\link[=layer_frame]{layer_frame()}} returns a list of two geoms, the
second a \code{\link[=layer_neatline]{layer_neatline()}} layer created using the frame object as the data
and the parameters \code{bgcolor = "none"} and \code{color = "none"}. asp is set to 1
if style is "circle" or "square" or the provided asp value otherwise.

Additional parameters passed through ... can include additional fixed
aesthetics (e.g. alpha). If using the fn parameter, the function is applied
to the frame simple feature object created by \code{\link[=make_frame]{make_frame()}} (not to the
original input data).
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))

raleigh_msa <-
  getdata::get_location(
    type = nc,
    name_col = "NAME",
    name = c("Franklin", "Johnston", "Wake"),
    crs = 3857
  )

ggplot() +
  layer_frame(
    data = raleigh_msa,
    frame = "circle",
    fill = "lightyellow",
    inscribed = FALSE
  ) +
  layer_location_data(
    data = raleigh_msa,
    mapping = aes(fill = NAME),
    alpha = 0.5
  ) +
  ggplot2::guides(
    fill = "none"
  )
}
\seealso{
Other layer: 
\code{\link{layer_location_data}()},
\code{\link{layer_neatline}()},
\code{\link{layer_scaled}()}
}
\concept{layer}
