% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_grouped.R
\name{layer_grouped}
\alias{layer_grouped}
\alias{layer_group_data}
\title{Make group layers}
\usage{
layer_grouped(
  data,
  mapping = NULL,
  groupname_col = "group",
  label_col = "name",
  geom = "sf",
  basemap = FALSE,
  palette = NULL,
  aesthetics = "fill",
  ...
)
}
\arguments{
\item{data}{Character string (e.g. url, file path, or name of data from
package) for a spatial data or a \code{sf}, \code{sfc}, or \code{bbox}  object with
geometry overlapping the location. If data is \code{NULL}, all unnamed
parameters are passed to \code{\link[sfext:read_sf_ext]{sfext::read_sf_ext()}} with a bbox based on
location. If data is not \code{NULL} and not a data.frame, url, file path, or
bbox, conversion to a sf object will still always be attempted with
\code{\link[sfext:as_sf]{sfext::as_sf()}}.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{groupname_col}{Group column name. Defaults to "group".}

\item{label_col}{Column name or id for a column with the text or labels to
pass to any text geom.}

\item{geom}{A character string indicating which ggplot2 geom to use, Default:
'sf'. Options include "sf" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}), "icon" (\code{\link[=layer_icon]{layer_icon()}}),
"markers" (\code{\link[=layer_markers]{layer_markers()}}), "sf_text" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_text()}}), and
"sf_label" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_label()}}). See details for a full list.}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. If \code{FALSE}, return
x as is.

If a \strong{ggplot}, add x to basemap object.

If a \strong{ggproto} object (or list that contains a \strong{ggproto} object), add x
and basemap object to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{palette}{Name of palette as a string. Must be on the form
packagename::palettename.}

\item{aesthetics}{Aesthetic to map to groupname_col. Defaults to "fill"; also
supports "color" or c("fill", "color").}

\item{...}{Additional parameters passed to \code{\link[=layer_location_data]{layer_location_data()}}}
}
\description{
Can be used to make multiple layers or multiple maps based on a grouping
variable.
}
\details{
Scales are applied a palette and aesthetic are provided and basemap is set to
\code{TRUE}.
}
