% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_grouped.R
\name{layer_grouped}
\alias{layer_grouped}
\alias{layer_group_data}
\title{Make group layers}
\usage{
layer_grouped(
  data,
  mapping = NULL,
  groupname_col = "group",
  label_col = "name",
  geom = "sf",
  basemap = FALSE,
  palette = NULL,
  aesthetics = "fill",
  ...
)
}
\arguments{
\item{data}{Character string (e.g. url, file path, or name of data from
package), a \code{sf}, \code{sfc}, or \code{bbox}  object including data in area.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{groupname_col}{Group column name}

\item{label_col}{Column name or id for a column with the text or labels to
pass to any text geom.}

\item{geom}{A character string indicating which ggplot2 geom to use, Default:
'sf'. Options include "sf" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}), "icon" (\code{\link[=layer_icon]{layer_icon()}}),
"markers" (\code{\link[=layer_markers]{layer_markers()}}), "sf_text" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_text()}}), and
"sf_label" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_label()}}). See details for a full list.}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \link[ggplot2:ggplot]{ggplot2::ggplot}. If \code{FALSE}, return x
as is.

If a \strong{ggplot}, add x to basemap object.}

\item{...}{Additional parameters passed to \link{layer_location_data}}

\item{layers}{If \code{TRUE}; return a list of layers; if \code{FALSE}; return a list
of ggplot2 maps; defaults to \code{TRUE}}
}
\description{
Can be used to make multiple layers or multiple maps based on a grouping
variable.
}
\details{
Scales are applied a palette and aesthetic are provided and basemap is set to
\code{TRUE}.
}
