% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_icon.R
\name{layer_icon}
\alias{layer_icon}
\alias{geom_sf_icon}
\title{Use ggsvg to add a layer with icons at feature locations}
\usage{
layer_icon(
  data = NULL,
  iconname_col = "icon",
  icon = NULL,
  px = NULL,
  source = NULL,
  svg = NULL,
  crs = getOption("maplayer.crs", default = 3857),
  ...
)

geom_sf_icon(
  data = NULL,
  iconname_col = "icon",
  icon = NULL,
  px = NULL,
  source = NULL,
  svg = NULL,
  crs = getOption("maplayer.crs", default = 3857),
  ...
)
}
\arguments{
\item{data}{A \code{sf} object. Any objects with polygon geometry are converted to
points using \code{\link[sf:geos_unary]{sf::st_centroid()}}.}

\item{iconname_col}{The column name in the input data to use as the icon
name. If the name matches multiple icons, the first match from \code{map_icons}
is used. You may provide a px or source value to select a different match
if needed but, in that case, all icons must use the same px or source
value. Note that the icon column should not be mapped with
\code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{icon}{Icon name. Default \code{NULL}. If \code{icon} is provided, \code{iconname_col}
is ignored. See \code{map_icons$name} for supported options.}

\item{px}{Icon size in pixels. See \code{map_icons$px} for supported options.
Optional but may be necessary to differentiate icons with duplicate names.}

\item{source}{Icon source. See \code{map_icons$repo} for supported options.
Optional but may be required to differentiate icons with duplicate names.}

\item{svg}{Optional. Custom file path or URL with SVG to pass to \code{svg}
parameter for  \code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}}.  If \code{icon} is provided, \code{svg} is
ignored.}

\item{crs}{Coordinate reference system; defaults to \code{NULL}.}

\item{...}{
  Arguments passed on to \code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg}}
  \describe{
    \item{\code{defaults}}{Advanced option.  A named list of default values for new aesthetics.
       In general this is not necessary when using \code{css()} aesthetics, as a
       default value will be determined based upon the CSS property e.g. \code{stroke}
       property will have a default value of "black"

Set `options(GGSVG_DEBUG = TRUE)` for some verbose debugging which will
cause ggsvg to output (to the console) the final SVG for each and every element
in the plot.}
  }}
}
\description{
Use the \code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}} function to plot icons using the centroids
from the input simple feature object to set the icon location.
}
\examples{
nc <- getdata::get_location(type = system.file("shape/nc.shp", package = "sf"), crs = 3857)

basemap <-
  ggplot2::ggplot() +
  ggplot2::theme_void() +
  layer_location_data(data = nc)

# icon can be set by name matching a name from map_icons
basemap +
  layer_icon(data = nc, icon = "point-start", size = 8)

# layer_icon can also use a column from the sf object
nc$icon <- rep(c("1", "2", "3", "4"), nrow(nc) / 4)

basemap +
  layer_icon(data = nc, iconname_col = "icon", size = 6)
}
\seealso{
\code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}}
\link{map_icons}
}
