% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_icon.R
\name{layer_icon}
\alias{layer_icon}
\alias{geom_sf_icon}
\title{Add a SVG icons to a ggplot2 plot based on a simple feature object}
\usage{
layer_icon(
  data = NULL,
  iconname_col = "icon",
  icon = NULL,
  px = NULL,
  source = NULL,
  svg = NULL,
  color = "black",
  crs = NULL,
  ...
)

geom_sf_icon(
  data = NULL,
  iconname_col = "icon",
  icon = NULL,
  px = NULL,
  source = NULL,
  svg = NULL,
  color = "black",
  crs = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{sf} object. Any objects with polygon geometry are converted to
points using \code{\link[sf:geos_unary]{sf::st_centroid()}}.}

\item{iconname_col}{The column name in the input data to use as the icon
name. If the name matches multiple icons, the first match from \code{map_icons}
is used. You may provide a px or source value to select a different match
if needed but, in that case, all icons must use the same px or source
value. Note that the icon column should not be mapped with
\code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{icon}{Icon name. Default \code{NULL}. If \code{icon} is provided, \code{iconname_col}
is not used.}

\item{px}{Icon size in pixels. See \code{map_icons$px} for supported options.
Optional but may be necessary to differentiate icons with duplicate names.}

\item{source}{Icon source. See \code{map_icons$repo} for supported options.
Optional but may be required to differentiate icons with duplicate names.}

\item{svg}{Optional. Custom file path or URL with SVG to pass to \code{svg}
parameter for  \code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}}.  If \code{icon} is provided, \code{svg} is
not used.}

\item{color}{SVG color passed to  \code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}}. default color is
set to "black".}

\item{crs}{Coordinate reference system; defaults to \code{NULL}.}

\item{...}{Additional parameters to  \code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}}.}
}
\description{
Use the  \code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}} function to plot icons using the centroids
from the input simple feature object to set the icon location.
}
\examples{
nc <- sfext::read_sf_path(system.file("shape/nc.shp", package = "sf"))

basemap <-
  ggplot2::ggplot() +
  ggplot2::theme_void() +
  layer_location_data(data = nc)

basemap +
  geom_sf_icon(data = nc, icon = "point-start", size = 10)

nc$icon <- rep(c("1", "2", "3", "4"), nrow(nc) / 4)

# FIXME: Not working - possibly some change in geom_point_svg?
# basemap +
#  layer_icon(data = nc, size = 5)
}
\seealso{
\code{\link[ggsvg:geom_point_svg]{ggsvg::geom_point_svg()}}
\link{map_icons}
}
