% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_inset.R
\name{layer_inset}
\alias{layer_inset}
\alias{make_inset_map}
\alias{stamp_inset_img}
\title{Use patchwork to create a map with an inset context map or figpatch to stamp
an inset image}
\usage{
layer_inset(
  inset,
  map = NULL,
  position = "bottomright",
  scale = 1,
  nudge_x = 0,
  nudge_y = 0,
  align_to = "full",
  ...
)

make_inset_map(
  inset = NULL,
  map = NULL,
  location = NULL,
  context = NULL,
  position = "bottomright",
  scale = 1,
  nudge_x = 0,
  nudge_y = 0,
  align_to = "full",
  ...
)

stamp_inset_img(
  plot,
  path,
  img_margin = ggplot2::margin(0, 0, 0, 0),
  position = "bottomright",
  scale = 1,
  nudge_x = 0,
  nudge_y = 0,
  align_to = "full",
  ...
)
}
\arguments{
\item{inset}{plot or map created with \link{ggplot2} passed to p argument of
\link[patchwork:inset_element]{patchwork::inset_element}. If both location and context are provided to
\link{make_inset_map}, inset is optional and any provided value is replaced with
a new layer created by \link{layer_location_context}.}

\item{position}{inset map position, Default: 'bottomright'. position,
nudge_x, and nudge_y are used to set the left, bottom, top, and right
parameters for \link[patchwork:inset_element]{patchwork::inset_element}.}

\item{scale}{scale of inset map, defaults to 1.}

\item{nudge_x, nudge_y}{nudge x and/or y position of inset map, Default: 0.}

\item{align_to}{Specifies what \code{left}, \code{bottom}, etc should be relative to.
Either \code{'panel'} (default), \code{'plot'}, or \code{'full'}.}

\item{...}{
  Arguments passed on to \code{\link[patchwork:inset_element]{patchwork::inset_element}}
  \describe{
    \item{\code{p}}{A grob, ggplot, patchwork, formula, raster, or nativeRaster object
to add as an inset}
    \item{\code{left,bottom,right,top}}{numerics or units giving the location of the
outer bounds. If given as numerics they will be converted to \code{npc} units.}
    \item{\code{on_top}}{Logical. Should the inset be placed on top of the other plot or
below (but above the background)?}
    \item{\code{clip}}{Logical. Should clipping be performed on the inset?}
    \item{\code{ignore_tag}}{Logical. Should autotagging ignore the inset?}
  }}

\item{context}{A sf object for context area or a ggplot layer representing
the context.}

\item{plot, map}{plot or map created with \link{ggplot2}}

\item{path}{image path passed to \link[figpatch:fig]{figpatch::fig} for \link{stamp_inset_img}}

\item{img_margin}{margin around image for \link{stamp_inset_img} created by
\link[ggplot2:element]{ggplot2::margin}. Defaults to no margin.}
}
\value{
ggplot2 map with inset map added using patchwork
}
\description{
\link{layer_inset_map} is useful when you want to add an inset to a plot.
}
\details{
\link{make_inset_map} is useful for creating an inset map just using the location
with fewer options for customization. In that case, the ... parameters are
passed to \link{layer_context} instead of \link[patchwork:inset_element]{patchwork::inset_element}

\link{stamp_inset_img} is useful for applying a logo to a map. The ... parameters
are passed to \link[figpatch:fig]{figpatch::fig}

Note, currently, plots created with \link{layer_inset} do not work with
\link{map_ggsave_ext} using the \code{single_file = TRUE} parameter.
}
