% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_labelled.R
\name{layer_labelled}
\alias{layer_labelled}
\alias{layer_label}
\alias{layer_show_label}
\title{Label simple feature objects in a location}
\usage{
layer_labelled(
  data,
  location = NULL,
  geom = "text",
  fn = NULL,
  label_col = NULL,
  mapping = NULL,
  union = FALSE,
  clip = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  drop_shadow = FALSE,
  shadow_params = NULL,
  ...
)
}
\arguments{
\item{data}{Data to use for labels (must be an sf object or a data frame that
can be converted to an sf object using \code{\link[sfext:as_sf]{sfext::as_sf()}})}

\item{location}{Location to label (if not specified the data is assumed to
conver the whole location);}

\item{geom}{A geom to use "text", "label", "textsf", "labelsf", "text_repel",
or "label_repel" or a geom function (passed to layer_fn).}

\item{fn}{Function to apply to data before creating labels; can be used in
the creation of the label_col.}

\item{label_col}{Label column name}

\item{mapping}{Aesthetic mapping, Default: \code{NULL}}

\item{union}{If TRUE, group by label_col and union geometry, Default: \code{FALSE}}

\item{clip}{Character string describing the part of the area to clip or
remove. Options include c("top", "right", "bottom", "left", "topright",
"bottomright", "bottomleft", "topleft"). If NULL, the area is not clipped
and a full edge can be returned.}

\item{dist}{Numeric. Distance to use for the edge. Default NULL
meters. Use negative values for an inside edge or positive numbers for an
outside edge.}

\item{diag_ratio}{Alternate way to define edge distance.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{drop_shadow}{If \code{TRUE}, use \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}} to add a drop shadow
to the label layer with shadow_params. Defaults to \code{FALSE}.}

\item{shadow_params}{Parameters passed to \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}} if \code{drop_shadow = TRUE}. Defaults to \code{list(x_offset = 5, y_offset = 5, sigma = 0.5)}.}

\item{...}{
  Arguments passed on to \code{\link[=layer_location_data]{layer_location_data}}
  \describe{
    \item{\code{smooth_params}}{Optional. Logical or a list of parameters passed to
\code{\link[smoothr:smooth]{smoothr::smooth()}}. If \code{TRUE}, apply \code{\link[smoothr:smooth]{smoothr::smooth()}} to location data
using default parameters. smooth_params is ignored if data is \code{NULL}
(inheriting data from ggplot).}
    \item{\code{asp}}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}
    \item{\code{package}}{Name of the package to search for data.}
    \item{\code{filetype}}{File type to use if passing parameters to
\code{\link[sfext:read_sf_ext]{sfext::read_sf_download()}} or \code{\link[sfext:read_sf_ext]{sfext::read_sf_pkg()}} (required for extdata
and cached data).}
    \item{\code{crop}}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}
    \item{\code{trim}}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}
    \item{\code{crs}}{Coordinate reference system to return.}
    \item{\code{basemap}}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. If \code{FALSE}, return x
as is.

If a \strong{ggplot}, add x to basemap object.

If a \strong{ggproto} object (or list that contains a \strong{ggproto} object), add x
and basemap object to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}
  }}
}
\description{
Label markers, streets, areas, or other simple feature objects using any of
the following geoms: "text", "sf_text", "label", "sf_label", "textsf",
"labelsf", "text_repel", or "label_repel".
}
\details{
Note: Unlike in some getdata package functions, fn is applied to all of the
data; not a subset of the data based on location. For this function, dist and
unit are both used by \code{\link[sfext:st_clip]{sfext::st_clip()}} (not by \code{\link[=layer_location_data]{layer_location_data()}})

The function also overrides the label aesthetics to hide the colored letters
that would otherwise appear when using a theme legend.
}
\seealso{
\code{\link[sfext:st_clip]{sfext::st_clip()}},\code{\link[=layer_location_data]{layer_location_data()}}
}
