% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_labelled.R
\name{layer_labelled}
\alias{layer_labelled}
\alias{layer_label}
\alias{layer_show_label}
\title{Label simple feature objects in a location}
\usage{
layer_labelled(
  data,
  location = NULL,
  geom = "text",
  fn = NULL,
  label_col = NULL,
  mapping = NULL,
  union = FALSE,
  clip = NULL,
  dist = NULL,
  unit = NULL,
  drop_shadow = FALSE,
  x_offset = 5,
  y_offset = 5,
  sigma = 0.5,
  ...
)
}
\arguments{
\item{data}{Data to use for label creation}

\item{location}{Location to label (if not specified the data is assumed to conver the whole location);}

\item{geom}{A geom to use "text", "label", "textsf", "labelsf", "text_repel", or "label_repel"}

\item{fn}{Function to apply to data before creating labels; can be used in the creation of the label_col}

\item{label_col}{Label column name}

\item{mapping}{Aesthetic mapping, Default: NULL}

\item{union}{If TRUE, group by label_col and union geometry, Default: \code{FALSE}}

\item{clip}{Character string describing the part of the area to clip or
remove. Options include c("top", "right", "bottom", "left", "topright",
"bottomright", "bottomleft", "topleft"). If NULL, the area is not clipped
and a full edge can be returned.}

\item{dist}{Numeric. Distance to use for the edge. Default NULL
meters. Use negative values for an inside edge or positive numbers for an
outside edge.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{drop_shadow}{If \code{TRUE}, use \link[ggfx:with_shadow]{ggfx::with_shadow} to add a drop shadow to the label layer. Defaults to \code{FALSE}.}

\item{x_offset, y_offset, sigma}{Parameters passed to ggfx::with_shadow if \code{drop_shadow = TRUE}.}

\item{...}{Additional parameters passed to \link[overedge:layer_location_data]{overedge::layer_location_data}}
}
\description{
Label markers, streets, areas, or other simple feature objects using any of
the following geoms: "text", "label", "textsf", "labelsf", "text_repel", or
"label_repel".
}
\details{
Note:
\itemize{
\item Unlike in some overedge package functions, fn is applied to the whole data; not a
subset of the data based on location.
\item For this function, dist and unit are both used by \code{\link[overedge:st_clip]{overedge::st_clip()}} (not by
\link{layer_location_data})
}

The function also overrides the label aesthetics to hide the colored letters
that would otherwise appear when using a theme legend.
}
\seealso{
\code{\link[overedge:st_clip]{overedge::st_clip()}},\code{\link[overedge:layer_location_data]{overedge::layer_location_data()}}
}
