% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_location_context.R
\name{layer_location_context}
\alias{layer_location_context}
\alias{layer_show_context}
\title{Create a map layer showing a location in context of another area or location}
\usage{
layer_location_context(
  data = NULL,
  fill = "gray70",
  color = "black",
  context = NULL,
  context_aes = list(fill = "white", color = "black", alpha = 1, ...),
  layer_between = NULL,
  crs = NULL,
  neatline = TRUE,
  ...
)
}
\arguments{
\item{data}{sf object with location data}

\item{fill}{Fill color for location data}

\item{color}{Edge color for location data}

\item{context}{sf or bbox object for context area or a \code{geom} layer}

\item{context_aes}{list with aesthetic attributes for context area; must
include fill and color; defaults to list(fill = "white", color = "black",
...)}

\item{layer_between}{An optional ggplot2 layer to insert between the context
layer and the location layer.}

\item{crs}{Coordinate reference system to return}

\item{neatline}{If \code{TRUE}, add a neatline layer to the returned layer}

\item{...}{Additional parameters passed to overedge::layer_location_data for
location}
}
\description{
Intended for use with inset_map function.
}
