% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_location_context.R
\name{layer_location_context}
\alias{layer_location_context}
\alias{layer_show_context}
\title{Create a layer showing a location and related context}
\usage{
layer_location_context(
  data = NULL,
  location = NULL,
  fill = "gray70",
  color = "black",
  context = NULL,
  context_params = list(fill = "white", color = "black", alpha = 1, ...),
  crs = getOption("maplayer.crs", default = 3857),
  mid_layer = NULL,
  neatline = TRUE,
  basemap = FALSE,
  ...
)
}
\arguments{
\item{data, location}{A data and location can be a either \code{sf} object or a
ggplot layer. Either data or location is required and, if data and location
are both provided, location is ignored. If data is \code{NULL} and location is a
formula or function, the context data is passed to the location function
and the results used as the data for the location layer.}

\item{fill, color}{Fill and color fixed aesthetics for location data.}

\item{context}{A \code{sf} object for context area or a ggplot layer representing
the context.}

\item{context_params}{A list with parameters for context layer; defaults to
\code{list(fill = "white", color = "black", alpha = 1, ...)}.}

\item{crs}{Coordinate reference system to return.}

\item{mid_layer}{A ggplot2 layer to insert between the context
layer and the location layer. Optional.}

\item{neatline}{A logical object, \code{CoordSf} object, or a list containing a
\code{CoordSf} object (typically from \code{\link[=layer_neatline]{layer_neatline()}}) added to layer.
\itemize{
\item If logical and \code{TRUE}, add a neatline layer using data from the context
layer with \code{color = NA} and \code{bgcolor = "none"}.
\item If object from \code{\link[=layer_neatline]{layer_neatline()}}, add it as is.
}}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. If \code{FALSE}, return
x as is.

If a \strong{ggplot}, add x to basemap object.

If a \strong{ggproto} object (or list that contains a \strong{ggproto} object), add x
and basemap object to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{...}{Additional parameters passed to \code{\link[=layer_location_data]{layer_location_data()}} if data or
location is an \code{sf} object.}
}
\description{
Create a ggplot2 layer for a location in a provided context. This function is
useful for making inset locator maps in combination with the \link{make_inset_map}
function.
}
