% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_location_data.R
\name{layer_location_data}
\alias{layer_location_data}
\title{Layer location data into a ggplot2 map}
\usage{
layer_location_data(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  location = NULL,
  dist = getOption("maplayer.dist"),
  diag_ratio = getOption("maplayer.diag_ratio"),
  unit = getOption("maplayer.unit", default = "meter"),
  asp = getOption("maplayer.asp"),
  package = getOption("maplayer.data_package"),
  filetype = getOption("maplayer.data_filetype"),
  fn = NULL,
  layer_fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = getOption("maplayer.from_crs", 4326),
  crs = getOption("maplayer.crs", 3857),
  label_col = "name",
  smooth_params = NULL,
  shadow_params = NULL,
  basemap = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{Character string (e.g. url, file path, or name of data from
package), a \code{sf}, \code{sfc}, or \code{bbox}  object including data in area.}

\item{geom}{A character string indicating which ggplot2 geom to use, Default:
'sf'. Options include "sf" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}), "icon" (\code{\link[=layer_icon]{layer_icon()}}),
"markers" (\code{\link[=layer_markers]{layer_markers()}}), "sf_text" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_text()}}), and
"sf_label" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_label()}}). See details for a full list.}

\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{unit to adjust location by dist or diag_ratio; defaults to
"meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{package}{Name of the package to search for data.}

\item{filetype}{File type to use if passing parameters to
\code{\link[sfext:read_sf_ext]{sfext::read_sf_download()}} or \code{\link[sfext:read_sf_ext]{sfext::read_sf_pkg()}} (required for extdata
and cached data).}

\item{fn}{Function to apply to data after filtering by location but before
returning from function.}

\item{layer_fn}{ggplot2 geom or custom function using lambda syntax. Use for
passing custom mapping functions to layer_location_data beyond the
supported geom options.}

\item{crop}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}

\item{from_crs}{Coordinate reference system used to match the location CRS to
the source data.}

\item{crs}{Coordinate reference system to return.}

\item{label_col}{Column name or id for a column with the text or labels to
pass to any text geom.}

\item{smooth_params}{Optional. Logical or a list of parameters passed to
\code{\link[smoothr:smooth]{smoothr::smooth()}}. If \code{TRUE}, apply \code{\link[smoothr:smooth]{smoothr::smooth()}} to location data
using default parameters. smooth_params is ignored if data is \code{NULL}
(inheriting data from ggplot).}

\item{shadow_params}{Optional. Logical or a list of parameters passed to
\code{\link[ggfx:with_shadow]{ggfx::with_shadow()}}. If \code{TRUE}, apply \code{\link[ggfx:with_shadow]{ggfx::with_shadow()}} to the layer
using default parameters. shadow_params is ignored if layer_fn is provided.}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. If \code{FALSE}, return x
as is.

If a \strong{ggplot}, add x to basemap object.

If a \strong{ggproto} object (or list that contains a \strong{ggproto} object), add x
and basemap object to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{...}{Additional parameters passed to selected geom or layer_fn}
}
\description{
Helper function to make a ggplot2 layer from data returned by
\code{\link[=get_location_data]{get_location_data()}}. For text geoms, the required aesthetic mapping is
set based on the name_col but values passed to mapping take precedence.
}
\section{Using the geom parameter}{


This function provides a convenient option for filtering data by location
while calling a different layer function from ggplot2, maplayer, or a
different package.

Options for the geom parameter from ggplot2:
\itemize{
\item "sf" (\link[ggplot2:ggsf]{ggplot2::geom_sf} - default)
\item "sf_text" or "text (\link[ggplot2:ggsf]{ggplot2::geom_sf_text})
\item "sf_label" or "label (\link[ggplot2:ggsf]{ggplot2::geom_sf_label})
}

Options for the geom parameter included in this package include:
\itemize{
\item "location" (\link{layer_location})
\item "context" or "location_context" (\link{layer_location_context})
\item "icon" (\link{layer_icon}),
\item "mapbox" (\link{layer_mapbox})
\item "markers" (\link{layer_markers})
\item "numbers" or "numbered" (\link{layer_numbers})
\item "mark" or "marked" (\link{layer_marked})
}

Options for the geom parameter from other suggested packages include:
\itemize{
\item "textsf" (\link[geomtextpath:geom_textsf]{geomtextpath::geom_textsf})
\item "labelsf" (\link[geomtextpath:geom_textsf]{geomtextpath::geom_labelsf})
\item "text_repel" (\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel})
\item "label_repel" (\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel})
\item "sf_pattern" or "pattern" (\link[ggpattern:geom-docs]{ggpattern::geom_sf_pattern})
}

\code{stat = "sf_coordinates"} is automatically added to the parameters for both
ggrepel functions. \code{label = .data[[name_col]]}  is automatically added to all
provided geoms where label is a required parameter.
}

\section{Using the layer_fn parameter}{


layer_fn can be a purrr-style lamba function (converted with
\code{\link[rlang:as_function]{rlang::as_function()}}) or a function.
}

\seealso{
Other layer: 
\code{\link{layer_frame}()},
\code{\link{layer_neatline}()},
\code{\link{layer_scaled}()}
}
\concept{layer}
