% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_mapbox.R
\name{layer_mapbox}
\alias{layer_mapbox}
\title{Use mapboxapi to make a Mapbox static map layer}
\usage{
layer_mapbox(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "meter",
  asp = NULL,
  style_url = "mapbox://styles/mapbox/satellite-streets-v11",
  style_id = NULL,
  username = NULL,
  basemap = FALSE,
  scale = 0.75,
  scaling_factor = "1x",
  attribution = TRUE,
  logo = TRUE,
  access_token = NULL,
  neatline = TRUE,
  color = "black",
  bgcolor = "white",
  size = 1,
  linetype = "solid",
  expand = FALSE,
  hide_grid = TRUE,
  label_axes = "----",
  ...
)
}
\arguments{
\item{data}{\code{sf}, \code{sfc}, or \code{bbox} object; any objects convertible with \link{as_bbox}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Buffer units; defaults to meter.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[overedge:get_asp]{get_asp()}} returns the same value
without modification.}

\item{style_url}{Map style url used to fill style_id and username parameters,
Default: "mapbox://styles/mapbox/satellite-streets-v11"}

\item{style_id}{A style ID (required if style_url is NULL).}

\item{username}{A Mapbox username (required if style_url is NULL).}

\item{basemap}{If FALSE, create a standalone layer; if \code{TRUE}, the layer is
precededed by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} to allow use as a basemap, Default: \code{TRUE}}

\item{scale}{ratio to scale the output image; `scale = 1` will return the
largest possible image. defaults to 0.5}

\item{scaling_factor}{The scaling factor of the tiles; either \code{"1x"}
(the default) or \code{"2x"}}

\item{attribution}{Controls whether there is attribution on the image.
Defaults to `TRUE`. Note: If attribution=false, the watermarked attribution
is removed from the image. You still have a legal responsibility to
attribute maps that use OpenStreetMap data, which includes most maps from
Mapbox. If you specify attribution=false, you are legally required to
include proper attribution elsewhere on the webpage or document.}

\item{logo}{Controls whether there is a Mapbox logo on the image. Defaults to
`TRUE`.}

\item{access_token}{A Mapbox access token; can be set with
\code{mb_access_token()}.}

\item{neatline}{If \code{TRUE}, add a neatline matching the provided data, Default:
\code{TRUE}}

\item{color}{Color of panel border, Default: 'black'}

\item{bgcolor}{Fill color of panel background; defaults to "white". If
"none", panel background is set to ggplot2::element_blank()}

\item{size}{Size of panel border, Default: 1}

\item{linetype}{Line type of panel border, Default: 'solid'}

\item{expand}{Default \code{FALSE.} If \code{TRUE}, the function adds
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} and
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} to expand the map extent to
provided parameters.}

\item{hide_grid}{If \code{TRUE}, hide major grid lines. Default: \code{TRUE}}

\item{label_axes}{A description of which axes to label passed to
\code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}}; defaults to '----' which hides axes
labels.}

\item{...}{Additional parameter passed to \link[mapboxapi:layer_static_mapbox]{mapboxapi::layer_static_mapbox}}
}
\description{
Use mapboxapi to make a Mapbox static map layer
}
\seealso{
\code{\link[mapboxapi:layer_static_mapbox]{mapboxapi::layer_static_mapbox()}}
}
