% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_markers.R
\name{layer_markers}
\alias{layer_markers}
\alias{layer_numbers}
\alias{make_markers}
\title{Create a ggplot2 layer with map markers or numbered markers}
\usage{
layer_markers(
  data,
  mapping = NULL,
  geom = "sf",
  make = FALSE,
  groupname_col = NULL,
  group_meta = NULL,
  crs = getOption("maplayer.crs", default = 3857),
  number = FALSE,
  num_by_group = FALSE,
  num_style = NULL,
  num_start = 1,
  suffix = NULL,
  sort = "dist_xmin_ymax",
  desc = FALSE,
  fn = NULL,
  ...
)

layer_numbers(
  data,
  mapping = NULL,
  geom = "label",
  make = FALSE,
  groupname_col = NULL,
  style = "roundrect",
  size = 5,
  sort = "dist_xmin_ymax",
  num_by_group = FALSE,
  num_style = NULL,
  num_start = 1,
  suffix = NULL,
  desc = FALSE,
  fn = NULL,
  crs = getOption("maplayer.crs", default = 3857),
  label.size = 0,
  label.padding = ggplot2::unit(size/10, "lines"),
  label.r = label.padding * 1.5,
  hjust = 0.5,
  vjust = 0.5,
  ...
)

make_markers(
  data,
  groupname_col = NULL,
  group_meta = NULL,
  join = sf::st_intersects,
  geo = FALSE,
  coords = c("lon", "lat"),
  address = "address",
  point = TRUE,
  crs = NULL,
  fn = NULL,
  ...
)
}
\arguments{
\item{data}{Character string (e.g. url, file path, or name of data from
package) for a spatial data or a \code{sf}, \code{sfc}, or \code{bbox}  object with
geometry overlapping the location. If data is \code{NULL}, all unnamed
parameters are passed to \code{\link[sfext:read_sf_ext]{sfext::read_sf_ext()}} with a bbox based on
location. If data is not \code{NULL} and not a data.frame, url, file path, or
bbox, conversion to a sf object will still always be attempted with
\code{\link[sfext:as_sf]{sfext::as_sf()}}.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{geom}{A character string indicating which ggplot2 geom to use, Default:
'sf'. Options include "sf" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}), "icon" (\code{\link[=layer_icon]{layer_icon()}}),
"markers" (\code{\link[=layer_markers]{layer_markers()}}), "sf_text" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_text()}}), and
"sf_label" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_label()}}). See details for a full list.}

\item{make}{If \code{TRUE}, pass data to \link{make_markers}.}

\item{groupname_col}{Group column name, used to join group metadata if
group_meta is a non-spatial data frame; Default: \code{NULL}}

\item{group_meta}{Group metadata as a data frame or sf object that intersect
with markers (using join function); Default: \code{NULL}}

\item{crs}{Coordinate reference system for markers, Default: \code{NULL}}

\item{number}{If \code{TRUE}, number markers using \code{\link[=layer_markers]{layer_markers()}} (not
currently supported)}

\item{num_by_group}{If \code{TRUE}, numbers are added by group based on
groupname_col.}

\item{num_style}{Style of enumeration, either "arabic", "alph", "Alph",
"roman", "Roman".}

\item{num_start}{Starting number; defaults to 1.}

\item{suffix}{Character to appended to "number" column. (e.g. "." for "1." or
":" for "1:"). Can also be a character vector with the same length as the
number column.}

\item{sort}{Sort column name, Default: "dist_xmin_ymax".}

\item{desc}{If \code{TRUE}, sort descending; default \code{FALSE}.}

\item{fn}{Function to apply to data before results; gives warning if data is
grouped; Default: \code{NULL}}

\item{...}{Additional parameters passed to \code{\link[=get_location_data]{get_location_data()}} when using
\code{make = TRUE} to pass data to \link{make_markers}}

\item{style}{Marker style; defaults to \code{NULL} for \code{\link[=layer_markers]{layer_markers()}} (supports
"facet"); defaults to "roundrect" for \code{\link[=layer_markers]{layer_markers()}} when numbered =
\code{TRUE} (default is only supported option at present).}

\item{size}{Marker size, Default: 5}

\item{label.size}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Replaced by the
\code{linewidth} aesthetic. Size of label border, in mm.}

\item{label.padding}{Amount of padding around label. Defaults to 0.25 lines.}

\item{label.r}{Radius of rounded corners. Defaults to 0.15 lines.}

\item{hjust, vjust}{Horizontal and vertical justification.}

\item{join}{Spatial relation function to combine data with group_meta, passed
to \code{\link[sf:st_join]{sf::st_join()}}. Defaults to \code{\link[sf:geos_binary_pred]{sf::st_intersects()}}.}

\item{geo}{If \code{FALSE}, pass data to \code{\link[getdata:get_location_data]{getdata::get_location_data()}} with \code{geo = TRUE} parameter.}

\item{coords}{Coordinate columns for input data.frame or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{address}{Address column name passed to \code{\link[tidygeocoder:geocode]{tidygeocoder::geocode()}} or
\link[tidygeocoder:geo]{tidygeocoder::geo}}

\item{point}{If \code{TRUE} and data does not have POINT or MULTIPOINT geometry,
convert to POINT data using \code{\link[sf:geos_unary]{sf::st_centroid()}}.}
}
\value{
ggplot2 layers
}
\description{
If make is \code{TRUE}, groupname_col, group_meta, crs, and fn is all passed on to
\link{make_markers}.
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))
nc <- sf::st_transform(nc, 3857)

basemap <-
  ggplot() +
  layer_location_data(
    data = nc,
    fill = NA
  )

basemap +
  layer_markers(
    data = nc[1:10],
    mapping = aes(size = AREA),
    make = TRUE
  )

large_nc <-
  getdata::get_location_data(
    data = nc,
    fn = ~ dplyr::filter(.x, AREA > 0.2)
  )

large_nc$number <- 1
large_nc$dist <- 2

basemap +
  layer_numbers(
    data = large_nc,
    mapping = aes(fill = NAME),
    sort = "dist_xmax_ymin",
    num_style = "Roman",
    geom = "label",
    size = 3
    ) +
  guides(fill = "none")

}
