% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_markers.R
\name{layer_markers}
\alias{layer_markers}
\alias{layer_numbers}
\title{Create a ggplot2 layer with map markers or numbered markers}
\usage{
layer_markers(
  data,
  mapping = NULL,
  geom = "sf",
  make = FALSE,
  groupname_col = NULL,
  group_meta = NULL,
  crs = NULL,
  number = FALSE,
  num_by_group = FALSE,
  num_style = NULL,
  num_start = 1,
  suffix = NULL,
  sort = "dist_xmin_ymax",
  desc = FALSE,
  fn = NULL,
  ...
)

layer_numbers(
  data,
  mapping = NULL,
  geom = "label",
  make = FALSE,
  groupname_col = NULL,
  style = "roundrect",
  size = 5,
  sort = "dist_xmin_ymax",
  num_by_group = FALSE,
  num_style = NULL,
  num_start = 1,
  suffix = NULL,
  desc = FALSE,
  fn = NULL,
  crs = NULL,
  label.size = 0,
  label.padding = ggplot2::unit(size/10, "lines"),
  label.r = label.padding * 1.5,
  hjust = 0.5,
  vjust = 0.5,
  ...
)
}
\arguments{
\item{data}{Data with markers, passed to data parameter of
\code{\link[overedge:get_location_data]{get_location_data()}}}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{geom}{The geometric object to use display the data}

\item{make}{If \code{TRUE}, pass data to \link{make_markers}.}

\item{groupname_col}{Group column name, used to join group metadata if
group_meta is a non-spatial data frame; Default: \code{NULL}}

\item{group_meta}{Group metadata as a data frame or sf object that intersect
with markers; Default: \code{NULL}}

\item{crs}{Coordinate reference system for markers, Default: \code{NULL}}

\item{number}{If \code{TRUE}, number markers using \code{\link[=layer_markers]{layer_markers()}} (not
currently supported)}

\item{num_by_group}{If \code{TRUE}, numbers are added by group based on
groupname_col.}

\item{num_style}{Style of enumeration, either "arabic", "alph", "Alph",
"roman", "Roman"}

\item{num_start}{Starting number; defaults to 1.}

\item{suffix}{Character to appended to "number" column. (e.g. "." for "1." or
":" for "1:"). Can also be a character vector with the same length as the
number column.}

\item{sort}{Sort column name, Default: "dist_xmin_ymax".}

\item{desc}{If \code{TRUE}, sort descending; default \code{FALSE}.}

\item{fn}{Function to apply to data before results; gives warning if data is
grouped; Default: \code{NULL}}

\item{...}{Additional parameters passed to \code{\link[=layer_location_data]{layer_location_data()}}}

\item{style}{Style of number markers to map; defaults to "roundrect".}

\item{size}{Marker size, Default: 5}

\item{label.size}{Size of label border, in mm.}

\item{label.padding}{Amount of padding around label. Defaults to 0.25 lines.}

\item{label.r}{Radius of rounded corners. Defaults to 0.15 lines.}

\item{hjust, vjust}{Horizontal and vertical justification.}
}
\value{
ggplot2 layers
}
\description{
If get is \code{TRUE}, groupname_col, group_meta, crs, and fn is all passed on to
\link[overedge:make_markers]{overedge::make_markers}.
}
\examples{
nc <- sfext::read_sf_path(system.file("shape/nc.shp", package = "sf"))
nc <- sfext::st_transform_ext(nc, 3857)

basemap <-
  ggplot2::ggplot() +
  layer_location_data(
    data = nc,
    fill = NA
  ) +
  ggplot2::theme_void()

basemap +
  layer_markers(
    data = nc[1:10],
    mapping = ggplot2::aes(size = AREA),
    make = TRUE
  )

large_nc <-
  getdata::get_location_data(
    data = nc,
    fn = ~ dplyr::filter(.x, AREA > 0.2)
  )

large_nc$number <- 1
large_nc$dist <- 2

basemap +
  layer_numbers(
    data = large_nc,
    mapping = ggplot2::aes(fill = NAME),
    sort = "dist_xmax_ymin",
    num_style = "Roman",
    geom = "label",
    size = 3
    ) +
  ggplot2::guides(fill = "none")

}
