% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_mask.R
\name{layer_mask}
\alias{layer_mask}
\alias{set_mask}
\title{Create a mask layer based on a simple feature object}
\usage{
layer_mask(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = getOption("maplayer.crs", default = 3857),
  fill = "white",
  color = NA,
  alpha = 0.5,
  mask = NULL,
  neatline = FALSE,
  expand = TRUE,
  ...
)

set_mask(x = NULL, mask = TRUE, data = NULL, crs = NULL, ...)
}
\arguments{
\item{data}{\code{sf}, \code{sfc}, or \code{bbox} object. If dist, diag_ratio, and/or asp
are provided, data is adjusted to set the boundaries of the mask. If data
is not provided, \code{mask} is required. If data is \code{NA}, mask is continuous
otherwise the data is erased from the mask area.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Coordinate reference system of bounding box to return; defaults to
\code{NULL} which maintains the crs of the input object.}

\item{fill}{mask fill color; defaults to "white"}

\item{color}{mask edge color; defaults to \code{NA}}

\item{alpha}{mask alpha/transparency; defaults to 0.5}

\item{mask}{A \code{sf}, \code{sfc}, or \code{bbox} object to define the mask area.
\code{diag_ratio}, \code{dist}, and \code{asp} parameters are ignored if a \code{mask} is
provided. defaults to \code{NULL}}

\item{neatline}{A logical object, \code{CoordSf} object, or a list containing a
\code{CoordSf} object (typically from \code{\link[=layer_neatline]{layer_neatline()}}) added to layer by
\code{\link[=set_neatline]{set_neatline()}}.
\itemize{
\item If logical and \code{TRUE}, add a neatline layer using data, crs and any
additional parameters passed to ... If logical and \code{FALSE}, return x as is.
\item If object from \code{\link[=layer_neatline]{layer_neatline()}}, add it as is.
}}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.
Giving a logical vector will separately control the expansion for the four
directions (top, left, bottom and right). The \code{expand} argument will be
recycled to length 4 if necessary. Alternatively, can be a named logical
vector to control a single direction, e.g. \code{expand = c(bottom = FALSE)}.}

\item{...}{Additional parameters to pass to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}}

\item{x}{A \code{sf}, \code{sfc}, \code{bbox}, \code{sfg}, \code{Raster}, \code{Spatial}, \code{Extent},
\code{numeric}, or \code{character} object (a place name passed to
\code{\link[osmdata:getbb]{osmdata::getbb()}}). See \code{\link[sfext:as_bbox]{as_bbox()}} for more details.}
}
\value{
\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} function.
}
\description{
Returns a mask for an area or areas as a simple feature object. \code{neatline = TRUE} only works for this layer if data is passed directly; not inherited.
}
