% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location_map.R
\name{make_location_map}
\alias{make_location_map}
\alias{make_social_map}
\alias{make_image_map}
\alias{make_layer_map}
\title{Make a ggplot map using \code{layer_location_data()}}
\usage{
make_location_map(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  data = NULL,
  crs = NULL,
  paper = NULL,
  width = NULL,
  height = NULL,
  units = "in",
  orientation = NULL,
  geom = "sf",
  basemap = TRUE,
  bg_layer = NULL,
  layer = NULL,
  fg_layer = NULL,
  addon = NULL,
  neatline = FALSE,
  labs_ext_params = list(...),
  save = FALSE,
  ggsave_params = list(dpi = 300, ...),
  ...,
  env = caller_env(),
  call = caller_env()
)

make_social_map(
  location,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = 3857,
  image = NULL,
  platform = NULL,
  format = NULL,
  orientation = NULL,
  basemap = TRUE,
  geom = "mapbox",
  save = FALSE,
  ggsave_params = list(fileext = "jpeg", dpi = 72, ...),
  ...
)

make_image_map(
  image_path,
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  data = NULL,
  crs = 3857,
  paper = "Letter",
  orientation = NULL,
  geom = "mapbox",
  style_url = NULL,
  basemap = TRUE,
  bg_layer = NULL,
  fg_layer = NULL,
  save = FALSE,
  ggsave_params = list(dpi = 300, ...),
  image_geom = "label",
  groupname_col = NULL,
  group_meta = NULL,
  number = FALSE,
  num_by_group = FALSE,
  num_style = NULL,
  num_start = 1,
  suffix = NULL,
  sort = "dist_xmin_ymax",
  desc = FALSE,
  ...
)

make_layer_map(
  bg_layer = NULL,
  layer = NULL,
  fg_layer = NULL,
  addon = NULL,
  basemap = NULL,
  neatline = NULL,
  labs_ext_params = NULL,
  save = FALSE,
  ggsave_params = list(width = 5, height = 4, unit = "in", dpi = 300),
  env = caller_env(),
  call = caller_env()
)
}
\arguments{
\item{location}{A sf object passed to \code{\link[=layer_location_data]{layer_location_data()}} with data if
layer is \code{NULL}. location is ignored if layer is provided. If data is
\code{NULL}, location is passed as data to facilitate using this function with
\code{geom = "mapbox"} where data is used to define the map area. Defaults to
\code{NULL}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{data}{Character string (e.g. url, file path, or name of data from
package) for a spatial data or a \code{sf}, \code{sfc}, or \code{bbox}  object with
geometry overlapping the location. If data is \code{NULL}, all unnamed
parameters are passed to \code{\link[sfext:read_sf_ext]{sfext::read_sf_ext()}} with a bbox based on
location. If data is not \code{NULL} and not a data.frame, url, file path, or
bbox, conversion to a sf object will still always be attempted with
\code{\link[sfext:as_sf]{sfext::as_sf()}}.}

\item{crs}{Coordinate reference system of bounding box to return; defaults to
\code{NULL} which maintains the crs of the input object.}

\item{paper}{Paper matching name from \code{paper_sizes} (e.g. "letter"). Not case
sensitive.}

\item{width}{Page width in "in", "px" or "mm" units. Default: \code{NULL}}

\item{height}{Page height in "in", "px" or "mm" units. Default: \code{NULL}}

\item{units}{Units to convert page dimensions to using \code{\link[papersize:convert_unit_type]{convert_unit_type()}}.}

\item{orientation}{Page orientation, Default: \code{NULL}. Supported options are
"portrait", "landscape", or "square".}

\item{geom}{A character string indicating which ggplot2 geom to use, Default:
'sf'. Options include "sf" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}), "icon" (\code{\link[=layer_icon]{layer_icon()}}),
"markers" (\code{\link[=layer_markers]{layer_markers()}}), "sf_text" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_text()}}), and
"sf_label" (\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_label()}}). See details for a full list.}

\item{basemap}{Either a logical vector or ggplot object.

If \strong{logical} and \code{TRUE}, add x to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. If \code{FALSE}, return
x as is.

If a \strong{ggplot}, add x to basemap object.

If a \strong{ggproto} object (or list that contains a \strong{ggproto} object), add x
and basemap object to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{bg_layer, fg_layer, addon}{A ggplot2 layer or a list of ggproto objects
(e.g. scales, labels, etc.) to add to the background or foreground of the
primary map layer defined by \code{"geom"} and other parameters. If the geom
creates an opaque layer or layer is an opaque layer (e.g. a layer produced
by \code{\link[=layer_mapbox]{layer_mapbox()}}) that covers the full map extent, the bg_layer will not
be visible.}

\item{layer}{A ggplot2 layer or a list of ggproto objects. If layer is
provided, all parameters passed to \code{\link[=layer_location_data]{layer_location_data()}} (including data,
location, dist, diag_ratio, unit, asp, crs, and geom) will be ignored. In
this case, the function simply stacks the bg_layer, layer, and fg_layer
objects then applies the basemap and neatline (using the \code{\link[=set_basemap]{set_basemap()}}
and \code{\link[=set_neatline]{set_neatline()}} helper functions.)}

\item{neatline}{A logical object, \code{CoordSf} object, or a list containing a
\code{CoordSf} object (typically from \code{\link[=layer_neatline]{layer_neatline()}}) added to layer by
\code{\link[=set_neatline]{set_neatline()}}.
\itemize{
\item If logical and \code{TRUE}, add a neatline layer using data, crs and any
additional parameters passed to ... If logical and \code{FALSE}, return x as is.
\item If object from \code{\link[=layer_neatline]{layer_neatline()}}, add it as is.
}}

\item{labs_ext_params}{Optional parameters passed to \code{\link[=labs_ext]{labs_ext()}}.}

\item{save}{If \code{TRUE}, save file with \link{ggsave_ext} using \code{ggsave_params}.
Defaults to \code{FALSE}.}

\item{ggsave_params}{List of parameters passed to \code{\link[papersize:ggsave_ext]{papersize::ggsave_ext()}}.}

\item{...}{Additional parameters passed to \code{\link[=layer_location_data]{layer_location_data()}} for
\code{\link[=make_location_map]{make_location_map()}} or \code{\link[=make_social_map]{make_social_map()}} or to \code{\link[=layer_markers]{layer_markers()}} for
\code{\link[=make_image_map]{make_image_map()}}.}

\item{env}{Environment for evaluation of \code{\link[=labs_ext]{labs_ext()}} if labs_ext_params is
supplied.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{image}{Image name passed to name parameter of \code{\link[papersize:get_social_size]{get_social_size()}}.}

\item{platform}{Social media platform, "Instagram", "Facebook", or "Twitter",
Default: \code{NULL}}

\item{format}{Image format, "post", "story", or "cover", Default: \code{NULL}}

\item{image_path}{path to location of images for \code{\link[=make_image_map]{make_image_map()}}}

\item{style_url}{A Mapbox style url; defaults to \code{NULL}.}

\item{image_geom}{For \code{\link[=make_image_map]{make_image_map()}}, geom to use with layer_markers to
mark the location of images (based on EXIF metadata).}

\item{groupname_col}{Group column name, used to join group metadata if
group_meta is a non-spatial data frame; Default: \code{NULL}}

\item{group_meta}{Group metadata as a data frame or sf object that intersect
with markers (using join function); Default: \code{NULL}}

\item{number}{If \code{TRUE}, number markers using \code{\link[=layer_markers]{layer_markers()}} (not
currently supported)}

\item{num_by_group}{If \code{TRUE}, numbers are added by group based on
groupname_col.}

\item{num_style}{Style of enumeration, either "arabic", "alph", "Alph",
"roman", "Roman".}

\item{num_start}{Starting number; defaults to 1.}

\item{suffix}{Character to appended to "number" column. (e.g. "." for "1." or
":" for "1:"). Can also be a character vector with the same length as the
number column.}

\item{sort}{Sort column name, Default: "dist_xmin_ymax".}

\item{desc}{If \code{TRUE}, sort descending; default \code{FALSE}.}
}
\description{
Location is used as the data parameter of layer_location_data so this
function is primarily appropriate for the layer_mapbox (\code{geom = "mapbox"}).
}
\details{
Using \code{\link[=make_image_map]{make_image_map()}}:

\code{\link[=make_image_map]{make_image_map()}} wraps \code{\link[sfext:read_sf_exif]{sfext::read_sf_exif()}} and \code{\link[=make_location_map]{make_location_map()}}.
It is designed for making simple maps of photos in combination with reference
tables.
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))

make_location_map(
  location = nc
)

name <- "North Carolina"

make_location_map(
  location = nc,
  labs_ext_params = list(
    title = "Map of {name}"
  )
)

make_location_map(
  data = nc,
  location = nc[2, ],
  dist = 2,
  unit = "mi",
  crop = FALSE,
  addon = ggplot2::theme_minimal(),
  labs_ext_params = list(
    title = "Map of {nc[2, ]$NAME} County"
  )
)

make_location_map(
  basemap = ggplot(data = nc[2, ]),
  fg_layer = geom_sf_text(aes(label = NAME)),
  data = nc,
  location = nc[2, ],
  mapping = aes(fill = NAME),
  addon = guides(fill = "none"),
  dist = 2,
  unit = "mi",
  crop = FALSE,
  neatline = TRUE,
  labs_ext_params = list(
    title = "Map of {nc[2, ]$NAME} and surrounding {name} counties"
  )
)
}
