% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mapbox_map.R
\name{make_mapbox_map}
\alias{make_mapbox_map}
\title{Make a map using layer_mapbox}
\usage{
make_mapbox_map(
  data = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "meter",
  asp = NULL,
  style_url = "mapbox://styles/mapbox/satellite-streets-v11",
  style_id = NULL,
  username = NULL,
  basemap = TRUE,
  scale = 0.75,
  scaling_factor = "1x",
  attribution = TRUE,
  logo = TRUE,
  access_token = NULL,
  neatline = NULL,
  color = "black",
  bgcolor = "white",
  linewidth = 0.5,
  linetype = "solid",
  expand = TRUE,
  hide_grid = TRUE,
  label_axes = "----",
  crs = 3857,
  ...
)
}
\arguments{
\item{data}{A \code{sf}, \code{sfc}, or \code{bbox} object.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{style_url}{Map style url used to fill style_id and username parameters,
Default: "mapbox://styles/mapbox/satellite-streets-v11"}

\item{style_id}{A style ID (required if style_url is \code{NULL}).}

\item{username}{A Mapbox username (required if \code{style_url = NULL}).}

\item{basemap}{If FALSE, create a standalone layer; if \code{TRUE}, the layer is
preceded by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} to allow use as a basemap, Default:
\code{TRUE}}

\item{scale}{ratio to scale the output image; \code{scale = 1} will return the
largest possible image. defaults to 0.5}

\item{scaling_factor}{The scaling factor of the tiles; either \code{"1x"}
(the default) or \code{"2x"}}

\item{attribution}{Controls whether there is attribution on the image.
Defaults to \code{TRUE}. If \code{FALSE}, the watermarked attribution is removed from
the image. You still have a legal responsibility to attribute maps that use
OpenStreetMap data, which includes most maps from Mapbox. If you specify
\code{attribution = FALSE}, you are legally required to include proper
attribution elsewhere on the webpage or document.}

\item{logo}{Controls whether there is a Mapbox logo on the image. Defaults to
\code{TRUE}.}

\item{access_token}{A Mapbox access token; which can be set with
\link[mapboxapi]{mb_access_token}.}

\item{neatline}{If \code{TRUE}, add a neatline matching the provided data,
Default: \code{TRUE}}

\item{color}{Color of panel border, Default: 'black'}

\item{bgcolor}{Fill color of panel background; defaults to "white". If
"none", panel background is set to \code{\link[ggplot2:element]{ggplot2::element_blank()}}}

\item{linewidth}{Line width of panel border, Default: 0.5}

\item{linetype}{Line type of panel border, Default: 'solid'}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{hide_grid}{If \code{TRUE}, hide grid lines. Default: \code{TRUE}}

\item{label_axes}{A description of which axes to label passed to
\code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}}; defaults to '----' which hides all axes labels.}

\item{crs}{Must be in web mercator projection.}

\item{...}{
  Arguments passed on to \code{\link[=make_layer_map]{make_layer_map}}
  \describe{
    \item{\code{labs_ext_params}}{Optional parameters passed to \code{\link[=labs_ext]{labs_ext()}}.}
    \item{\code{ggsave_params}}{List of parameters passed to \code{\link[papersize:ggsave_ext]{papersize::ggsave_ext()}}.}
    \item{\code{layer}}{A ggplot2 layer or a list of ggproto objects. If layer is
provided, all parameters passed to \code{\link[=layer_location_data]{layer_location_data()}} (including data,
location, dist, diag_ratio, unit, asp, crs, and geom) will be ignored. In
this case, the function simply stacks the bg_layer, layer, and fg_layer
objects then applies the basemap and neatline (using the \code{\link[=set_basemap]{set_basemap()}}
and \code{\link[=set_neatline]{set_neatline()}} helper functions.)}
    \item{\code{save}}{If \code{TRUE}, save file with \link{ggsave_ext} or \link{ggsave_social},
requires \code{basemap = TRUE} and filename is not \code{NULL} \emph{or} \code{...} must
include a name parameter. Defaults to \code{FALSE}.}
    \item{\code{env}}{Environment for evaluation of \code{\link[=labs_ext]{labs_ext()}} if labs_ext_params is
supplied.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
  }}
}
\description{
Wraps \code{\link[=make_layer_map]{make_layer_map()}} and passes layer created with \code{\link[=layer_mapbox]{layer_mapbox()}} to
bg_layer and the neatline parameters to \code{\link[=layer_neatline]{layer_neatline()}} (using the same
data as the mapbox background layer). The neatline parameters are only used
if neatline is NULL.
}
\keyword{internal}
