% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplayer-package.R
\docType{package}
\name{maplayer-package}
\alias{maplayer}
\alias{maplayer-package}
\title{maplayer: Make ggplot2 Map Layers}
\description{
Make map-making with ggplot2 and sf more convenient with a consistent set of wrapper functions. Relies on the overedge package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/maplayer}
  \item \url{https://elipousson.github.io/maplayer/}
  \item Report bugs at \url{https://github.com/elipousson/maplayer/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
