% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot}
\alias{aes}
\alias{last_plot}
\alias{margin}
\alias{guides}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{guides}}, \code{\link[ggplot2]{last_plot}}, \code{\link[ggplot2:element]{margin}}}
}}

