% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-overedge.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{layer_location_data}
\alias{layer_neatline}
\alias{st_buffer_ext}
\alias{st_bbox_ext}
\alias{st_transform_ext}
\alias{as_sf}
\alias{as_sf_list}
\alias{is_sf}
\alias{is_sf_list}
\alias{is_point}
\alias{is_multipoint}
\alias{is_polygon}
\alias{is_multipolygon}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{overedge}{\code{\link[overedge]{as_sf}}, \code{\link[overedge:as_sf]{as_sf_list}}, \code{\link[overedge:is_geom_type]{is_multipoint}}, \code{\link[overedge:is_geom_type]{is_multipolygon}}, \code{\link[overedge:is_geom_type]{is_point}}, \code{\link[overedge:is_geom_type]{is_polygon}}, \code{\link[overedge]{is_sf}}, \code{\link[overedge:is_sf]{is_sf_list}}, \code{\link[overedge]{layer_location_data}}, \code{\link[overedge]{layer_neatline}}, \code{\link[overedge]{st_bbox_ext}}, \code{\link[overedge]{st_buffer_ext}}, \code{\link[overedge]{st_transform_ext}}}
}}

